+++ types/index.d.ts
@@ -1,6 +1,6 @@
 export { duration } from "./duration.js";
 export { tzOffset } from "./timezone.js";
-export { setTzSeparator } from "./utils/config.js";
 export { DateTime } from "./datetime-class.js";
 export { datetime, utc, datetimeTz } from "./datetime.js";
+export { setTimeSeparator, setTzSeparator } from "./config/index.js";
 export { daysBetween, weekNumber } from "./utils/date.js";